+++ src/tightdb/group_shared.hpp
@@ -217,8 +217,21 @@ public:
     /// group. Doing so will result in undefined behavior.
     void reserve(std::size_t size_in_bytes);
 
-    // Has db been modified since last transaction?
+    // Querying for changes:
+    //
+    // NOTE:
+    // "changed" means that one or more commits has been made to the database
+    // since the SharedGroup (on which wait_for_change() is called) last
+    // started, committed, promoted or advanced a transaction.
+    //
+    // No distinction is made between changes done by another process
+    // and changes done by another thread in the same process as the caller.
+    //
+    // Has db been changed ?
     bool has_changed();
+
+    // The calling thread goes to sleep until the database is changed.
+    void wait_for_change();
 
     // Transactions: