+++ test/test_shared.cpp
@@ -1788,6 +1788,83 @@ TEST_IF(Shared_AsyncMultiprocess, allow_async)
 #endif
 }
 
+static const int num_threads = 3;
+static int shared_state[num_threads];
+static Mutex muu;
+void waiter(string path, int i)
+{
+    SharedGroup sg(path, true, SharedGroup::durability_Full);
+    {
+        LockGuard l(muu);
+        shared_state[i] = 1;
+    }
+    sg.wait_for_change();
+    {
+        LockGuard l(muu);
+        shared_state[i] = 2; // this state should not be observed by the writer
+    }
+    sg.wait_for_change(); // we'll fall right through here, because we haven't advanced our readlock
+    {
+        LockGuard l(muu);
+        shared_state[i] = 3;
+    }
+    sg.begin_read();
+    sg.end_read();
+    sg.wait_for_change(); // this time we'll wait because state hasn't advanced since we did.
+    {
+        LockGuard l(muu);
+        shared_state[i] = 4;
+    }
+    // works within a read transaction as well
+    sg.begin_read();
+    sg.wait_for_change();
+    sg.end_read();
+    {
+        LockGuard l(muu);
+        shared_state[i] = 5;
+    }
+}
+
+TEST(Shared_WaitForChange)
+{
+    SHARED_GROUP_TEST_PATH(path);
+    for (int j=0; j < num_threads; j++)
+        shared_state[j] = 0;
+    SharedGroup sg(path, false, SharedGroup::durability_Full);
+    Thread threads[num_threads];
+    for (int j=0; j < num_threads; j++)
+        threads[j].start(bind(&waiter, string(path), j));
+    sleep(1);
+    for (int j=0; j < num_threads; j++) {
+        LockGuard l(muu);
+        CHECK_EQUAL(1, shared_state[j]);
+    }
+
+    sg.begin_write();
+    sg.commit();
+    sleep(1);
+    for (int j=0; j < num_threads; j++) {
+        LockGuard l(muu);
+        CHECK_EQUAL(3, shared_state[j]);
+    }
+    sg.begin_write();
+    sg.commit();
+    sleep(1);
+    for (int j=0; j < num_threads; j++) {
+        LockGuard l(muu);
+        CHECK_EQUAL(4, shared_state[j]);
+    }
+    sg.begin_write();
+    sg.commit();
+    sleep(1);
+    for (int j=0; j < num_threads; j++) {
+        LockGuard l(muu);
+        CHECK_EQUAL(5, shared_state[j]);
+    }
+    for (int j=0; j < num_threads; j++)
+        threads[j].join();
+}
+
 #endif // endif not on windows
 
 TEST(Shared_MixedWithNonShared)